STYLE: Improve attribution by ignoring bulk formatting
This file lists revisions that should be ignored when considering
attribution for the actual code written.  Code style changes should
not be considered as modifications with regards to attribution.

To see clean and meaningful blame information.
$ git blame important.py --ignore-revs-file .git-blame-ignore-revs

To configure git to automatically ignore revisions listed in a file on
every call to git blame.
$ git config blame.ignoreRevsFile .git-blame-ignore-revs